assistant/pyproject.toml
[project]
name = "assistant"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Glitch752", email = "xxGlitch752xx@gmail.com" }
]
dependencies = [
    "elevenlabs>=1.59.0",
    "python-dotenv>=1.1.0",
    "anthropic>=0.51.0",
    "pytz>=2025.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"assistant" = "assistant:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]
